apps/api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
pydantic-settings
# AI / Extraction
pydantic-ai[google]
openai
tiktoken
python-dotenv
aiofiles
pypdf
